Move single size malloc to tests
